requirements.txt
flask
flask-socketio
eventlet
opencv-python-headless
mediapipe
tensorflow
joblib
Pillow
numpy
